added hook for PQL Query logic
